wasm: fix initial window paint event
We were requesting to activate the window too soon, when the window was
still in the initial size of 600x600, in which case, the paint event was
not getting called. This would require the user to have to click the
window in order to see it.

Removal of this does not seem to effect showing of tooltips, which it
was added initially.

Change-Id: I3efa86be99a3d9e1d9eda87b7537fff1e96e6774
Fixes: QTBUG-102005
Reviewed-by: Aleksandr Reviakin <aleksandr.reviakin@qt.io>
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>
Reviewed-by: David Skoland <david.skoland@qt.io>
(cherry picked from commit e70c953a599a9dfb0d43bdea1772eb47b2d46ec6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>