Fix qmake build on FreeBSD
In Makefile.unix we have one target with multiple dependencies:
'qlibraryinfo_final.o'.  The macro $< is expanded to the first
dependency in GNU Make.  In BSD Make, this macro expands to nothing
if there's more than one dependency.

Fix this by using the explicit path to the source file for this one
target.

Fixes: QTBUG-65425
Change-Id: Ib51f5f181b670e5533d22ce546fce9a20a649a27
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>