Android: fix A11Y object description
This patch amends a374d59abc415eee1866322176b7762158f48abd.

That commit introduced refactoring of A11Y object description
announcements that also took a value into account.
However for the elements without values (like QPushButton), an unused
space is added at the end of the description.

This patch fixes the issue by explicitly checking that the value
interface exists and that the value string is not empty.

Fixes: QTBUG-102744
Pick-to: 6.3 6.2 5.15
Change-Id: Ic1ba50859fb91c871c242189967dcce35723a0b2
Reviewed-by: Andreas Buhr <andreas.buhr@qt.io>