QIOSTextResponder: recreate responder upon reset
Apple introduced UITextInteraction in iOS 13, which
caused issues with our text handling, and resulted in
e00d888daefb. This, again, seems to be the reason why
UIKit in some cases simply removes the input delegate
from our text responder. This typically happens after
a QPlatformInputContext::reset(), and then when the
next character is typed on the input panel (and hence,
leaves no informative stack trace).

The result of removing the input delegate, which is the
interface that let's Qt communicate IM changes back to
UIKit, leads Qt and the UIKit out of IM sync. E.g for a
japanese keyboard, after a QPlatformInputContext::reset(),
the Qt input field would remove the marked text, but
UIKit would still show japanese characters
based on what used to be marked text.

To work around this issue, this patch will therefore
recreate the text responder when Qt tells us to reset.
This will cause the first responder to change, which will
apparently also reset the UIKit IM state.

Fixes: QTBUG-102960
Pick-to: 6.3 6.2 5.15
Change-Id: I00901e980874fae819cc7d89a68fa34ae44808c2
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>